refactor(client/Hero/*): Fix/update responsibe para tamaño tablet y celular, eliminacion de código repetido, css actualizado
