+++ src/components/Hero/hero.module.css
@@ -1,0 +1,162 @@
+.hero_container {
+  width: 100%;
+  display: flex;
+  gap: 1rem;
+  align-items: flex-start;
+}
+
+.heroTop {
+  /* background-image: url('../../img/Vector\ 2.svg'); */
+  /* background-image: url('../../../public/vectores/Vector\ 2.svg');
+  background-repeat: no-repeat;
+  background-position: center; */
+}
+
+.heroMiddle {
+  /* background-image: url('../../../public/vectores/Vector\ 4.svg');
+  background-repeat: no-repeat;
+  background-position: center; */
+}
+
+.heroBottom {
+  margin-bottom: 7rem;
+  /* background-image: url('../../../public/vectores/Vector\ 6.svg');
+  background-repeat: no-repeat;
+  background-position: center; */
+}
+
+.heroText {
+  width: 85%;
+  margin: 0 auto;
+  margin-bottom: 10rem;
+  align-items: center;
+}
+
+.hero h2 {
+  margin: 5rem 0 2rem 1rem;
+  font-family: var(--font-main);
+  font-size: 2.875rem;
+  font-weight: 700;
+  color: transparent;
+  background: var(--color-title);
+  background-clip: text;
+  -webkit-background-clip: text;
+  -webkit-text-fill-color: transparent;
+}
+
+.hero span {
+  color: transparent;
+  background: var(--color-font);
+  background-clip: text;
+  -webkit-background-clip: text;
+  -webkit-text-fill-color: transparent;
+}
+
+.hero p {
+  color: var(--color-font);
+  margin-left: 1.25rem;
+  width: 80%;
+  font-size: 1.5rem;
+  font-weight: 600;
+  line-height: 3rem;
+  letter-spacing: 0.0375rem;
+}
+
+.hero .heroP {
+  width: 100%;
+  text-align: center;
+  font-size: 1.875rem;
+}
+
+@media screen and (max-width: 910px) {
+  .heroBottom {
+    flex-direction: column;
+    align-items: flex-start;
+    background-image: none;
+  }
+
+  .heroText {
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    background-image: none;
+    width: 100%;
+  }
+
+  .hero {
+    width: 60%;
+  }
+
+  .hero_img {
+    width: 40%;
+    -o-object-fit: contain;
+    object-fit: contain;
+  }
+
+  .hero_imgBottom {
+    width: 60%;
+  }
+
+  .hero_img img {
+    width: 100%;
+    margin-top: 5rem;
+  }
+
+  .hero_img .imgAcostado {
+    width: 100%;
+    margin-top: 8rem;
+    margin-left: 60%;
+    margin-bottom: -3rem;
+  }
+
+  .hero h2 {
+    font-size: 1.5rem;
+  }
+
+  .hero p {
+    margin-top: 1.125rem;
+  }
+
+  .hero .heroPMiddle,
+  .hero .heroPBottom {
+    line-height: 2.25rem;
+    width: 100%;
+    z-index: 3;
+    margin-top: 1.125rem;
+  }
+
+
+}
+
+@media screen and (max-width: 520px) {
+  .hero p {
+    font-size: 1rem;
+    line-height: 2.25rem;
+    width: 110%;
+    z-index: 3;
+    margin-left: -2rem;
+  }
+
+  .hero .heroPMiddle,
+  .hero .heroPBottom {
+    margin-left: 1rem;
+
+  }
+
+  .hero_img .imgParado {
+    margin-top: 16rem;
+  }
+
+  .heroText {
+    width: 100%;
+  }
+
+  .hero .heroP h2 {
+    font-size: 1.4rem;
+  }
+
+  .hero .heroP p {
+    font-size: 1rem;
+  }
+
+}
