--- src/components/Hero/heroTexto.module.css
@@ -1,50 +1,0 @@
-.hero_container {
-  width: 85%;
-  margin: 0 auto;
-  margin-bottom: 10rem;
-  display: flex;
-  gap: 1rem;
-  align-items: center;
-}
-
-
-
-.hero h2{
-  margin: 5rem 0 2rem 1rem;
-  font-family: var(--font-main);
-  font-size: 2.875rem;
-  font-weight: 700;
-  color: transparent;
-  background: var(--color-title);
-  background-clip: text;
-  -webkit-background-clip: text;
-  -webkit-text-fill-color: transparent;
-}
-.hero span {
-  color: transparent;
-  background: var(--color-font);
-  background-clip: text;
-  -webkit-background-clip: text;
-  -webkit-text-fill-color: transparent;
-
-}
-
-.hero p {
-  color: var(--color-font);
-  margin-left: 1.25rem;
-  
-  text-align: center;
-  font-size: 1.875rem;
-  font-weight: 600;
-  line-height: 3rem;
-  letter-spacing: 0.0375rem; 
-}
-
-@media screen and (max-width: 910px) {
-  .hero_container {
-    flex-direction: column;
-    align-items: center;
-    background-image: none;
-  }
-
-}
